Redid comments for mapbox style, separate out dashboardcontrol
