+++ app/scripts/components/exploration/components/map/tour-control.tsx
@@ -1,0 +1,45 @@
+import React, {useCallback} from 'react';
+import { useAtomValue } from 'jotai';
+import { useTour } from '@reactour/tour';
+import { CollecticonCircleQuestion } from '@devseed-ui/collecticons';
+
+import { SelectorButton } from '$components/common/map/style/button';
+import useThemedControl from '$components/common/map/controls/hooks/use-themed-control';
+import { introTourSteps } from '$components/exploration/tour-manager';
+import { timelineDatasetsAtom } from '$components/exploration/atoms/datasets';
+
+export function TourButtonComponent({onClick, disabled}: {
+  onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
+  disabled: boolean;
+}) {
+  return (
+  <SelectorButton
+    tipContent='Open guided tour'
+    tipProps={{ placement: 'left' }}
+    disabled={disabled}
+    onClick={onClick}
+  >
+    <CollecticonCircleQuestion />
+  </SelectorButton>);
+}
+
+export function TourInvokingControl(props) {
+  useThemedControl(() => <TourButtonComponent {...props} />, {
+    position: 'top-right'
+  });
+  return null;
+}
+
+export function TourDashboardControl() {
+  const { setIsOpen, setCurrentStep, setSteps } = useTour();
+
+  const reopenTour = useCallback(() => {
+    setCurrentStep(0);
+    setSteps?.(introTourSteps);
+    setIsOpen(true);
+  },[setIsOpen]);
+
+  const datasets = useAtomValue(timelineDatasetsAtom);
+  const disabled = datasets.length === 0;
+  return <TourInvokingControl onClick={reopenTour} disabled={disabled} />;
+}
